requirements.txt
alembic==1.0.10
amqp==2.4.2
aniso8601==6.0.0
appdirs==1.4.3
attrs==19.1.0
billiard==3.6.0.0
black==19.3b0
blinker==1.4
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
chrononaut==0.2.4
Click==7.0
Faker==1.0.7
Flask==1.0.3
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-Redis==0.3.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
future==0.17.1
idna==2.8
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
kombu==4.5.0
Mako==1.0.10
MarkupSafe==1.1.1
marshmallow==3.0.0rc4
parso==0.4.0
pip-autoremove==0.9.1
pluggy==0.11.0
psycopg2==2.8.2
PyJWT==1.7.1
pyls-black==0.4.4
PySocks==1.7.0
python-dateutil==2.8.0
python-dotenv==0.10.2
python-editor==1.0.4
python-jsonrpc-server==0.1.2
python-language-server==0.26.1
python-slugify==3.0.2
pytz==2019.1
redis==3.2.1
requests==2.22.0
sentry-sdk==0.8.0
six==1.12.0
SQLAlchemy==1.3.3
stripe==2.29.0
text-unidecode==1.2
toml==0.10.0
twilio==6.27.0
urllib3==1.25.3
vine==1.3.0
Werkzeug==0.15.4
WTForms==2.2.1
